+++ src/views/IncidentalsView/OngoingIncidentalsTable/CreateOngoingIncidentalsModal/CreateOngoingIncidentalsModal.tsx
@@ -4,7 +4,6 @@ import { v4 as uuid } from 'uuid';
 import { ValidationConstraint } from '../../../../utils/validation/constraints';
 import Validator from '../../../../utils/validation/validator';
 import DeductionTypeSelect from '_/components/form/DeductionTypeSelect/DeductionTypeSelect';
-import NumberTextField from '_/components/form/NumberTextField/NumberTextField';
 import GenericModal from '_/components/generic/GenericModal/GenericModal';
 import useFormValidation from '_/hooks/useFormValidation/useFormValidation';
 import useIncidentalsState from '_/hooks/useIncidentalsState/useIncidentalsState';
@@ -35,11 +34,6 @@ interface CreateOngoingIncidentalsInput {
    * Deduction type of the new incidentals
    */
   deductionType: DeductionType;
-
-  /**
-   * Interval in which the new incidentals need to be payed
-   */
-  invoiceInterval: number;
 }
 
 /**
@@ -59,12 +53,10 @@ function CreateOngoingIncidentalsModal(
   } = useFormValidation<CreateOngoingIncidentalsInput>({
     formValidator: new Validator<CreateOngoingIncidentalsInput>({
       name: ValidationConstraint.NoEmptyString,
-      invoiceInterval: ValidationConstraint.Defined,
     }),
     defaultFormInput: {
       name: '',
       deductionType: DeductionType.PerApartment,
-      invoiceInterval: undefined,
     },
     onSubmitSuccess: (values) => {
       addOngoingIncidentals({
@@ -72,7 +64,6 @@ function CreateOngoingIncidentalsModal(
         name: values.name,
         costs: [],
         deductionType: values.deductionType,
-        invoiceInterval: values.invoiceInterval,
       });
       props.onCloseModal();
     },
@@ -103,23 +94,10 @@ function CreateOngoingIncidentalsModal(
             helperText={formErrors.name}
           />
         </Grid>
-        <Grid item xs={6}>
+        <Grid item xs={12}>
           <DeductionTypeSelect
             value={formInput.deductionType as DeductionType}
             onChange={formInputSetters.deductionType}
-          />
-        </Grid>
-        <Grid item xs={6}>
-          <NumberTextField
-            required
-            label="Abrechnungszeitraum"
-            id="invoiceInterval"
-            value={formInput.invoiceInterval}
-            onChange={formInputSetters.invoiceInterval}
-            errorMessage={formErrors.invoiceInterval}
-            min={1}
-            max={12}
-            onlyInteger
           />
         </Grid>
       </Grid>